Isolate TokenSpeed transport policy
