+++ model_gateway/src/routers/grpc/multimodal/assembly/tokenspeed/transport.rs
@@ -1,0 +1,185 @@
+use std::sync::OnceLock;
+
+use llm_multimodal::Modality;
+use tracing::{info, warn};
+
+use crate::routers::grpc::{
+    context::WorkerSelection,
+    proto_wrapper::{
+        tokenspeed_mm_shm_min_bytes, tokenspeed_mm_tensor_transport_mode,
+        tokenspeed_shm_dev_writable,
+    },
+};
+
+pub(super) fn tokenspeed_encoder_input_dtype(
+    modality: Modality,
+    workers: Option<&WorkerSelection>,
+) -> String {
+    if let Some(dtype) = tokenspeed_encoder_input_dtype_from_env(modality) {
+        return dtype;
+    }
+    if let Some(dtype) = tokenspeed_encoder_input_dtype_from_worker(workers) {
+        return dtype;
+    }
+    "float32".to_string()
+}
+
+fn tokenspeed_encoder_input_dtype_from_env(modality: Modality) -> Option<String> {
+    static IMAGE_DTYPE: OnceLock<Option<String>> = OnceLock::new();
+    static VIDEO_DTYPE: OnceLock<Option<String>> = OnceLock::new();
+    static AUDIO_DTYPE: OnceLock<Option<String>> = OnceLock::new();
+    static DEFAULT_DTYPE: OnceLock<Option<String>> = OnceLock::new();
+
+    let modality_dtype = match modality {
+        Modality::Image | Modality::ImageEmbeds => {
+            cached_env_dtype(&IMAGE_DTYPE, "SMG_TOKENSPEED_IMAGE_ENCODER_INPUT_DTYPE")
+        }
+        Modality::Video => {
+            cached_env_dtype(&VIDEO_DTYPE, "SMG_TOKENSPEED_VIDEO_ENCODER_INPUT_DTYPE")
+        }
+        Modality::Audio => {
+            cached_env_dtype(&AUDIO_DTYPE, "SMG_TOKENSPEED_AUDIO_ENCODER_INPUT_DTYPE")
+        }
+    };
+    modality_dtype
+        .or_else(|| cached_env_dtype(&DEFAULT_DTYPE, "SMG_TOKENSPEED_ENCODER_INPUT_DTYPE"))
+}
+
+fn cached_env_dtype(cell: &'static OnceLock<Option<String>>, name: &str) -> Option<String> {
+    cell.get_or_init(|| std::env::var(name).ok().filter(|dtype| !dtype.is_empty()))
+        .clone()
+}
+
+fn tokenspeed_encoder_input_dtype_from_worker(workers: Option<&WorkerSelection>) -> Option<String> {
+    let worker = match workers? {
+        WorkerSelection::Single { worker } => worker,
+        WorkerSelection::Dual { prefill, .. } => prefill,
+    };
+    worker
+        .metadata()
+        .spec
+        .labels
+        .get("multimodal_encoder_dtype")
+        .filter(|dtype| !dtype.is_empty())
+        .cloned()
+}
+
+/// Resolve whether large multimodal tensors should use the SHM transport for
+/// this request. `shm` = always (legacy explicit opt-in); `auto` = only when the
+/// worker is known to share SMG's `/dev/shm`; anything else (including unset or
+/// `inline`) keeps the inline gRPC path.
+pub(super) fn resolve_tokenspeed_shm_enabled(
+    modality: Modality,
+    workers: Option<&WorkerSelection>,
+) -> bool {
+    let configured_mode = tokenspeed_mm_tensor_transport_mode();
+    let mode = effective_tokenspeed_transport_mode(modality, &configured_mode);
+    log_tokenspeed_transport_config_once(&configured_mode, &mode, modality);
+    match mode.as_str() {
+        // SHM only ever happens when SMG can actually write /dev/shm.
+        "shm" => tokenspeed_shm_dev_writable(),
+        "auto" => worker_shares_dev_shm(workers) && tokenspeed_shm_dev_writable(),
+        "" | "inline" => false,
+        other => {
+            log_unknown_tokenspeed_transport_once(other);
+            false
+        }
+    }
+}
+
+pub(in crate::routers::grpc::multimodal) fn effective_tokenspeed_transport_mode(
+    modality: Modality,
+    configured_mode: &str,
+) -> String {
+    if !configured_mode.is_empty() {
+        return configured_mode.to_string();
+    }
+
+    match modality {
+        Modality::Video => "auto".to_string(),
+        Modality::Image | Modality::ImageEmbeds | Modality::Audio => "inline".to_string(),
+    }
+}
+
+fn log_tokenspeed_transport_config_once(
+    configured_mode: &str,
+    effective_mode: &str,
+    modality: Modality,
+) {
+    static LOGGED: OnceLock<()> = OnceLock::new();
+    LOGGED.get_or_init(|| {
+        info!(
+            configured_mode,
+            effective_mode,
+            ?modality,
+            shm_min_bytes = tokenspeed_mm_shm_min_bytes(),
+            dev_writable = tokenspeed_shm_dev_writable(),
+            "TokenSpeed multimodal tensor transport configured"
+        );
+    });
+}
+
+fn log_unknown_tokenspeed_transport_once(value: &str) {
+    static WARNED: OnceLock<()> = OnceLock::new();
+    WARNED.get_or_init(|| {
+        warn!(
+            value,
+            "Unknown SMG_TOKENSPEED_MM_TENSOR_TRANSPORT value; expected inline|shm|auto, using inline"
+        );
+    });
+}
+
+/// Whether the worker is *verified* to share SMG's `/dev/shm`, making the SHM
+/// transport safe under `auto`.
+///
+/// Rather than inferring locality from the worker URL (TCP loopback proves only
+/// network locality, not a shared `/dev/shm`), the worker advertises its
+/// `/dev/shm` filesystem identity (`<boot_id>:<st_dev of /dev/shm>`) via
+/// `GetServerInfo`, which discovery stores in the worker's `shm_namespace_id`
+/// label. Two processes share `/dev/shm` iff these tokens match: `boot_id` pins
+/// the host, and `st_dev` is the tmpfs superblock device, identical whenever the
+/// same tmpfs backs both `/dev/shm` mounts — including separate containers that
+/// share it via `--ipc`/bind-mount (where mount-namespace inodes differ but the
+/// underlying superblock is the same). We compare the worker's token to ours:
+/// equal ⇒ shared. A missing/empty token or any mismatch is treated as
+/// non-sharing, so `auto` safely falls back to inline.
+fn worker_shares_dev_shm(workers: Option<&WorkerSelection>) -> bool {
+    let Some(local) = local_shm_namespace_id() else {
+        return false;
+    };
+    let worker = match workers {
+        Some(WorkerSelection::Single { worker }) => worker,
+        Some(WorkerSelection::Dual { prefill, .. }) => prefill,
+        None => return false,
+    };
+    worker
+        .metadata()
+        .spec
+        .labels
+        .get("shm_namespace_id")
+        .is_some_and(|id| !id.is_empty() && id == local)
+}
+
+/// This process's `/dev/shm` filesystem identity: `<boot_id>:<st_dev of /dev/shm>`.
+/// `boot_id` pins the host (it is not namespaced) and `st_dev` is the tmpfs
+/// superblock device backing `/dev/shm`; together they identify the tmpfs so two
+/// processes sharing it (even across containers via `--ipc`/bind-mount) produce
+/// the same token. Computed once; `None` if it can't be determined (then `auto`
+/// stays inline).
+pub(in crate::routers::grpc::multimodal) fn local_shm_namespace_id() -> Option<&'static str> {
+    static ID: OnceLock<Option<String>> = OnceLock::new();
+    ID.get_or_init(compute_shm_namespace_id).as_deref()
+}
+
+#[cfg(unix)]
+fn compute_shm_namespace_id() -> Option<String> {
+    use std::os::unix::fs::MetadataExt;
+    let boot_id = std::fs::read_to_string("/proc/sys/kernel/random/boot_id").ok()?;
+    let shm_dev = std::fs::metadata("/dev/shm").ok()?.dev();
+    Some(format!("{}:{shm_dev}", boot_id.trim()))
+}
+
+#[cfg(not(unix))]
+fn compute_shm_namespace_id() -> Option<String> {
+    None
+}
